[CL-2438] Move survey utils out of form builder and add goback to config
