+++ front/app/components/FormBuilder/utils.ts
@@ -1,10 +1,7 @@
 // Services
 import { IFlatCustomField } from 'services/formCustomFields';
-import { IPhaseData, updatePhase } from 'services/phases';
-import { IProjectData, updateProject } from 'services/projects';
-
-// Typings
-import { Multiloc } from 'typings';
+import { IPhaseData } from 'services/phases';
+import { IProjectData } from 'services/projects';
 
 // Utils
 import { isNilOrError } from 'utils/helperUtils';
@@ -13,60 +10,7 @@ export type FormBuilderConfig = {
   formBuilderTitle: { id: string; defaultMessage: string };
   formCustomFields: IFlatCustomField[] | undefined | Error;
   showStatusBadge: boolean;
-};
-
-type FormActionsConfig = {
-  phaseId?: string;
-  editFormLink: string;
-  viewFormLink: string;
-  viewFormResults: string;
-  heading?: Multiloc;
-  postingEnabled: boolean;
-  togglePostingEnabled: () => void;
-};
-
-const getSurveyPhases = (phases: IPhaseData[] | Error | null | undefined) => {
-  return !isNilOrError(phases)
-    ? phases.filter(
-        (phase) => phase.attributes.participation_method === 'native_survey'
-      )
-    : [];
-};
-
-export const getFormActionsConfig = (
-  project: IProjectData,
-  phases?: IPhaseData[] | Error | null | undefined
-): FormActionsConfig[] => {
-  const processType = project.attributes.process_type;
-  if (processType === 'continuous') {
-    return [
-      {
-        editFormLink: `/admin/projects/${project.id}/native-survey/edit`,
-        viewFormLink: `/projects/${project.attributes.slug}/ideas/new`,
-        viewFormResults: `/admin/projects/${project.id}/native-survey/results`,
-        postingEnabled: project.attributes.posting_enabled,
-        togglePostingEnabled: () => {
-          updateProject(project.id, {
-            posting_enabled: !project.attributes.posting_enabled,
-          });
-        },
-      },
-    ];
-  }
-
-  return getSurveyPhases(phases).map((phase) => ({
-    phaseId: phase.id,
-    editFormLink: `/admin/projects/${project.id}/phases/${phase.id}/native-survey/edit`,
-    viewFormLink: `/projects/${project.attributes.slug}/ideas/new?phase_id=${phase.id}`,
-    viewFormResults: `/admin/projects/${project.id}/native-survey/results?phase_id=${phase.id}`,
-    heading: phase.attributes.title_multiloc,
-    postingEnabled: phase.attributes.posting_enabled,
-    togglePostingEnabled: () => {
-      updatePhase(phase.id, {
-        posting_enabled: !phase.attributes.posting_enabled,
-      });
-    },
-  }));
+  getGoBackUrl: (projectId: string) => string;
 };
 
 export const getIsPostingEnabled = (
