+++ front/app/containers/Admin/projects/project/nativeSurvey/FormResults/CompletionBar.tsx
